move CLI commands to the commands directory
