+++ packages/patrol_cli/lib/src/command_runner.dart
@@ -7,13 +7,15 @@ import 'package:cli_completion/cli_completion.dart';
 import 'package:dispose_scope/dispose_scope.dart';
 import 'package:file/file.dart';
 import 'package:file/local.dart';
+import 'package:patrol_cli/src/commands/devices_command.dart';
+import 'package:patrol_cli/src/commands/doctor_command.dart';
+import 'package:patrol_cli/src/commands/test_command.dart';
+import 'package:patrol_cli/src/commands/update_command.dart';
 import 'package:patrol_cli/src/common/constants.dart';
 import 'package:patrol_cli/src/common/logger.dart';
 import 'package:patrol_cli/src/common/logging_local_process_manager.dart';
 import 'package:patrol_cli/src/common/tool_exit.dart';
 import 'package:patrol_cli/src/features/devices/device_finder.dart';
-import 'package:patrol_cli/src/features/devices/devices_command.dart';
-import 'package:patrol_cli/src/features/doctor/doctor_command.dart';
 import 'package:patrol_cli/src/features/run_commons/dart_defines_reader.dart';
 import 'package:patrol_cli/src/features/run_commons/test_finder.dart';
 import 'package:patrol_cli/src/features/test/android_test_backend.dart';
@@ -21,8 +23,6 @@ import 'package:patrol_cli/src/features/test/ios_deploy.dart';
 import 'package:patrol_cli/src/features/test/ios_test_backend.dart';
 import 'package:patrol_cli/src/features/test/native_test_runner.dart';
 import 'package:patrol_cli/src/features/test/pubspec_reader.dart';
-import 'package:patrol_cli/src/features/test/test_command.dart';
-import 'package:patrol_cli/src/features/update/update_command.dart';
 import 'package:platform/platform.dart';
 import 'package:process/process.dart';
 import 'package:pub_updater/pub_updater.dart';
