+++ packages/patrol_cli/lib/src/commands/devices_command.dart
