+++ packages/patrol_cli/lib/src/commands/doctor_command.dart
