+++ packages/patrol_cli/lib/src/commands/test_command.freezed.dart
