+++ packages/patrol_cli/lib/src/commands/update_command.dart
